+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 * Cap initial followed account retrieval before merging toots to timeline at 1,600; beyond that they will be loaded in a background job
 * Add `config.apibackgroundLoadSleepBetweenRequestsMS` param; sleep that long between requests during background loads
 * Fix `OUTLIER_DAMPENER` value in various presets
-* Sort filter options by displayName if available
-* ~~Fix removing newly muted accounts from feed~~
+* Fix removing newly muted accounts from feed
+* (Demo App) Sort filter options by 1displayName1 if available
 
 ## v0.68.0
 * Update the cache in the background instead of always dropping/reloading the whole thing when it gets stale